% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verification.R
\name{baseProps_ex1}
\alias{baseProps_ex1}
\title{USGS verification example: base parameterisation}
\usage{
baseProps_ex1()
}
\description{
USGS verification example
}
\references{
p.23, \url{https://pubs.usgs.gov/sir/2010/5102/support/sir2010-5102.pdf}
}
