% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verification.R
\name{compareModelResults}
\alias{compareModelResults}
\title{USGS verification example: compare R results for water level increase with other
models and calculate statistical goodness of fit values (e.g. RMSE, PBIAS, NSE)}
\usage{
compareModelResults(conf = example1())
}
\arguments{
\item{conf}{list as retrieved by example1()}
}
\value{
data.frame with R results and other model including all goodness of fit
criteria calculated with gof() of package hydrogof
}
\description{
USGS verification example: compare R results for water level increase with other
models and calculate statistical goodness of fit values (e.g. RMSE, PBIAS, NSE)
}
\references{
Table 5, p.25, \url{https://pubs.usgs.gov/sir/2010/5102/support/sir2010-5102.pdf}
}
\seealso{
\code{\link{example1}} for the USGS example parameterisation with distances
}
