% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/verification.R
\name{conf_ex1}
\alias{conf_ex1}
\title{USGS verification example: model parameterisation (one distance)}
\usage{
conf_ex1(x = 0, y = 0, baseProps = baseProps_ex1(), dbg = FALSE)
}
\arguments{
\item{x}{distance from the center of the recharge basin in the x direction (L)
(Default: 0 feet)}

\item{y}{distance from the center of the recharge basin in the y direction (L)
(Default: 0 feet)}

\item{baseProps}{basic model properties as retrieved by baseProps_ex1()}

\item{dbg}{If True additional messages on debug messages}
}
\description{
USGS verification example: model parameterisation (one distance)
}
\references{
p.23, http://pubs.usgs.gov/sir/2010/5102/support/sir2010-5102.pdf
}
\seealso{
\code{\link{baseProps_ex1}} for the USGS example parameterisation
}

