% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kwb.hantush.R
\name{erf}
\alias{erf}
\title{Error function}
\usage{
erf(x)
}
\arguments{
\item{x}{x}
}
\value{
Error function result
}
\description{
Error function
}
\references{
\url{https://stat.ethz.ch/R-manual/R-devel/library/stats/html/Normal.html}
}
