% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/verification.R
\name{getModelComparisonTable}
\alias{getModelComparisonTable}
\title{USGS verification example: get model comparison table}
\usage{
getModelComparisonTable()
}
\value{
data.frame with water level increase of different model approaches
}
\description{
USGS verification example: get model comparison table
}
\examples{
modelComparison <- getModelComparisonTable()
}
\references{
Table 5, p.25, http://pubs.usgs.gov/sir/2010/5102/support/sir2010-5102.pdf
}

