% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kwb.hantush.R
\name{hantushDistances}
\alias{hantushDistances}
\title{Hantush distance: for multiple coordinates}
\usage{
hantushDistances(x = 0:10, y = rep(0, length(x)),
  baseProps = baseProperties(), dbg = TRUE)
}
\arguments{
\item{x}{vector with distances from the center of the recharge basin in the x
direction (L), (Default: each meter between 0-100m)}

\item{y}{vector with distances from the center of the recharge basin in the y
direction (L), (Default: 0 times length of x)}

\item{baseProps}{basic model properties as retrieved by baseProperties()}

\item{dbg}{If True additional debug messages are printed on screen}
}
\value{
Head at a given time after recharge begins
}
\description{
Hantush distance: for multiple coordinates
}
\seealso{
\code{\link{hantush}} for parameterizing the Hantush equation
}
