% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kwb.hantush.R
\name{hantushS}
\alias{hantushS}
\title{Helper function hantushS}
\usage{
hantushS(x, alpha, beta)
}
\arguments{
\item{x}{distance between 0 and half lenght of recharge basin}

\item{alpha}{alpha}

\item{beta}{beta}
}
\value{
Hantush star
}
\description{
Helper function hantushS
}
\references{
p.22, http://pubs.usgs.gov/sir/2010/5102/support/sir2010-5102.pdf
}

