% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verification.R
\name{plotModelComparison}
\alias{plotModelComparison}
\title{USGS verification example: plot model comparison results}
\usage{
plotModelComparison(modelComparison = compareModelResults(),
  title = "", ...)
}
\arguments{
\item{modelComparison}{data.frame as retrieved by compareModelResults(),
Default: (compareModelResults())}

\item{title}{to be used as title above plot (Default: "")}

\item{...}{further arguments passed to function xyplot()}
}
\value{
model comparison goodness of fit plots
}
\description{
USGS verification example: plot model comparison results
}
\examples{
### Plot model comparison with title "Model comparison" and one plot for
### each page
plotModelComparison(
  title = "Model comparison",
  layout = c(1, 1)
)
}
\seealso{
\code{\link{compareModelResults}} for comparison with USGS benchmark models
}
