% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/verification.R
\name{compareModelResults}
\alias{compareModelResults}
\title{USGS verification example: compare R results for water level increase with other
models and calculate statitical goodness of fit values (e.g. RMSE, PBIAS, NSE)}
\usage{
compareModelResults(conf = confDistances_ex1())
}
\arguments{
\item{conf}{list as retrieved by confDistances_ex1()}
}
\value{
data.frame with R results and other model including all goodness of fit
criteria calculated with gof() of package hydrogof
}
\description{
USGS verification example: compare R results for water level increase with other
models and calculate statitical goodness of fit values (e.g. RMSE, PBIAS, NSE)
}
\references{
Table 5, p.25, http://pubs.usgs.gov/sir/2010/5102/support/sir2010-5102.pdf
}
\seealso{
\code{\link{confDistances_ex1}} for the USGS example parameterisation with distances
}

