% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kwb.hantush.R
\name{hantushSstar}
\alias{hantushSstar}
\title{Hantush function Sstar}
\usage{
hantushSstar(alpha, beta, dbg)
}
\arguments{
\item{alpha}{alpha}

\item{beta}{beta}

\item{dbg}{If True additional messages on integration quality of function
hantushSstar are printed on screen}
}
\value{
Hantush Sstar result
}
\description{
Hantush function Sstar
}
\references{
p.22, http://pubs.usgs.gov/sir/2010/5102/support/sir2010-5102.pdf
}

